All code refactor (I wanna die)
I found that checking all the nodes one by one to see which one was not a good idea. So my goal now is to make the abstract syntax tree run itself through a domino effect. One node executes the other, which executes the other and the other so on, from a function derived from the main node (luv u polymorphism).